+++ SmartHome/IphoneControllers/OtherTableViewCell.h
@@ -18,7 +18,7 @@
 @property(nonatomic, strong)NSString * deviceid;
 @property (nonatomic,weak) NSString *sceneid;
 //房间id
-@property (nonatomic,assign) NSInteger roomID;
+@property (nonatomic,assign) int roomID;
 @property (strong, nonatomic) Scene *scene;
 @property (nonatomic, assign) id<OtherTableViewCellDelegate>delegate;
